solve: 217 contains duplicate

+++ 02_remove_ele.cpp
@@ -6,7 +6,7 @@ class Solution{
 public:
     int removeElement(vector<int>& nums, int val){
         int k = 0;
-        for ( int i = 0; i < nums.size(); i++){
+        for ( int i = 0; i < (int)nums.size(); i++){
             if (nums[i] != val){
                 nums[k] = nums[i];
                 k++;
